Typescriptへ置き換え close #1
