+++ .env
@@ -1,7 +1,7 @@
-REACT_APP_FIREBASE_APIKEY="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-REACT_APP_FIREBASE_DOMAIN="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-REACT_APP_FIREBASE_DATABASE="https://XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-REACT_APP_FIREBASE_PROJECT_ID="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-REACT_APP_FIREBASE_STORAGE_BUCKET="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-REACT_APP_FIREBASE_SENDER_ID="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
-REACT_APP_FIREBASE_APP_ID="XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
+REACT_APP_FIREBASE_APIKEY="AIzaSyAsshOdheRFL8x0guuPmkak4RrdkepszOY"
+REACT_APP_FIREBASE_DOMAIN="todoapp-90ad5.firebaseapp.com"
+REACT_APP_FIREBASE_DATABASE="https://todoapp-90ad5.firebaseio.com"
+REACT_APP_FIREBASE_PROJECT_ID="todoapp-90ad5"
+REACT_APP_FIREBASE_STORAGE_BUCKET="todoapp-90ad5.appspot.com"
+REACT_APP_FIREBASE_SENDER_ID="272580573244"
+REACT_APP_FIREBASE_APP_ID="1:272580573244:web:0821b2938545383c7a2b39"
