+++ src/components/App.tsx
@@ -1,33 +1,37 @@
 /* eslint-disable jsx-a11y/no-redundant-roles */
-import React, { useState, useRef, useEffect } from "react";
-import Form from "./components/Form";
-import FilterButton from "./components/FilterButton";
-import Todo from "./components/Todo";
-import { db } from "./Firebase";
+import React, { useState, useRef, useEffect } from 'react';
+import Form from './Form';
+import FilterButton from './FilterButton';
+import Todo from './Todo';
+import { db } from '../Firebase';
 
-function usePrevious(value) {
-  const ref = useRef();
+function usePrevious(value: number) {
+  const ref = React.useRef(0);
   useEffect(() => {
     ref.current = value;
   });
   return ref.current;
 }
 
-const FILTER_MAP = {
+// All,Active,Completed
+// as const readonly unionと同じ扱い
+const FILTER_KEYS = ['All','Active','Completed'] as const;
+
+type FILTER_UNION = typeof FILTER_KEYS[number];
+
+const FILTER_MAP:{[FILTER_KEYS in FILTER_UNION]: any} = {
   All: () => true,
-  Active: (task) => !task.completed,
-  Completed: (task) => task.completed,
+  Active: (task: { completed: boolean; }) => !task.completed,
+  Completed: (task: { completed: boolean; }) => task.completed,
 };
 
-// All,Active,Completed
-const FILTER_NAMES = Object.keys(FILTER_MAP);
-
-function App(props) {
-  const [tasks, setTasks] = useState([{ id: "", name: "", completed: "" }]);
+function App() {
+  const [tasks, setTasks] = useState([{ id: '', name: '', completed: true }]);
   // 初期表示時はAll
-  const [filter, setFilter] = useState("All");
+  // stringでなくunionでないとだめ'All'|'Active'|'Completed'
+  const [filter, setFilter] = useState<'All'|'Active'|'Completed'>('All');
   useEffect(() => {
-    const unSub = db.collection("Tasks").onSnapshot((snapshot) => {
+    const unSub = db.collection('Tasks').onSnapshot((snapshot) => {
       setTasks(
         snapshot.docs.map((doc) => ({
           id: doc.id,
@@ -39,11 +43,11 @@ function App(props) {
     return () => unSub();
   }, []);
 
-  function toggleTaskCompleted(id) {
+  function toggleTaskCompleted(id:string) {
     const updatedTasks = tasks.map((task) => {
       // 編集されたタスクと同じIDを持っているIDを持っているか
       if (id === task.id) {
-        db.collection("Tasks")
+        db.collection('Tasks')
           .doc(id)
           .set({ completed: !task.completed }, { merge: true });
       }
@@ -52,15 +56,15 @@ function App(props) {
     setTasks(updatedTasks);
   }
 
-  function deleteTask(id) {
-    db.collection("Tasks").doc(id).delete();
+  function deleteTask(id:string) {
+    db.collection('Tasks').doc(id).delete();
   }
 
-  function editTask(id, newName) {
+  function editTask(id:string, newName:string) {
     const editedTaskList = tasks.map((task) => {
       // 編集されたタスクと同じIDを持っている場合
       if (id === task.id) {
-        db.collection("Tasks").doc(id).set({ name: newName }, { merge: true });
+        db.collection('Tasks').doc(id).set({ name: newName }, { merge: true });
       }
       return task;
     });
@@ -81,7 +85,7 @@ function App(props) {
       />
     ));
 
-  const filterList = FILTER_NAMES.map((name) => (
+  const filterList = FILTER_KEYS.map((name) => (
     <FilterButton
       key={name}
       name={name}
@@ -90,22 +94,23 @@ function App(props) {
     />
   ));
 
-  function addTask(name) {
-    db.collection("Tasks").add({
+  function addTask(name:string) {
+    db.collection('Tasks').add({
       name: name,
       completed: false,
     });
   }
 
-  const tasksNoun = taskList.length !== 1 ? "tasks" : "task";
+  const tasksNoun = taskList.length !== 1 ? 'tasks' : 'task';
   const headingText = `${taskList.length} ${tasksNoun} remaining`;
 
   const listHeadingRef = useRef(null);
   const prevTaskLength = usePrevious(tasks.length);
-
   useEffect(() => {
-    if (tasks.length - prevTaskLength === -1) {
-      listHeadingRef.current.focus();
+    if (prevTaskLength) {
+      if (tasks.length - prevTaskLength === -1) {
+          listHeadingRef.current.focus();
+      }
     }
   }, [tasks.length, prevTaskLength]);
 
@@ -113,7 +118,7 @@ function App(props) {
     <div className="todoapp stack-large">
       <Form addTask={addTask} />
       <div className="filters btn-group stack-exception">{filterList}</div>
-      <h2 id="list-heading" tabIndex="-1" ref={listHeadingRef}>
+      <h2 id="list-heading" ref={listHeadingRef}>
         {headingText}
       </h2>
       <ul
