+++ src/components/Todo.tsx
@@ -1,0 +1,121 @@
+import React, { useEffect, useRef, useState } from 'react';
+
+function usePrevious(value :boolean) {
+  const ref = useRef(false);
+  useEffect(() => {
+    ref.current = value;
+  });
+  return ref.current;
+}
+
+type todoProps = {
+  id:string;
+  name:string;
+  completed:boolean;
+  key:string;
+  toggleTaskCompleted(id:string):void
+  deleteTask(id:string):void
+  editTask(id: string,newName: string):void;
+}
+
+export default function Todo(props:todoProps) {
+  const [isEditing, setEditing] = useState(false);
+  const [newName, setNewName] = useState('');
+
+  const editFieldRef = useRef(null);
+  const editButtonRef = useRef(null);
+
+  const wasEditing = usePrevious(isEditing);
+
+  function handleChange(event: React.ChangeEvent<HTMLInputElement>) {
+    setNewName(event.target.value);
+  }
+
+  function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
+    event.preventDefault();
+    if (!newName.trim()) {
+      return;
+    }
+    props.editTask(props.id, newName);
+    setNewName('');
+    setEditing(false);
+  }
+
+  // 更新中の場合
+  const editingTemplate = (
+    <form className="stack-small" onSubmit={handleSubmit}>
+      <div className="form-group">
+        <label className="todo-label" htmlFor={props.id}>
+          New name for {props.name}
+        </label>
+        <input
+          id={props.id}
+          className="todo-text"
+          type="text"
+          value={newName || props.name}
+          onChange={handleChange}
+          ref={editFieldRef}
+        />
+      </div>
+      <div className="btn-group">
+        <button
+          type="button"
+          className="btn todo-cancel"
+          onClick={() => setEditing(false)}
+        >
+          Cancel
+          <span className="visually-hidden">renaming {props.name}</span>
+        </button>
+        <button type="submit" className="btn btn__primary todo-edit">
+          Save
+          <span className="visually-hidden">new name for {props.name}</span>
+        </button>
+      </div>
+    </form>
+  );
+
+  // 更新中でない場合
+  const viewTemplate = (
+    <div className="stack-small">
+      <div className="c-cb">
+        <input
+          id={props.id}
+          type="checkbox"
+          defaultChecked={props.completed}
+          onChange={() => props.toggleTaskCompleted(props.id)}
+        />
+        <label className="todo-label" htmlFor={props.id}>
+          {props.name}
+        </label>
+      </div>
+      <div className="btn-group">
+        <button
+          type="button"
+          className="btn"
+          onClick={() => setEditing(true)}
+          ref={editButtonRef}
+        >
+          Edit <span className="visually-hidden">{props.name}</span>
+        </button>
+        <button
+          type="button"
+          className="btn btn__danger"
+          onClick={() => props.deleteTask(props.id)}
+        >
+          Delete <span className="visually-hidden">{props.name}</span>
+        </button>
+      </div>
+    </div>
+  );
+
+  useEffect(() => {
+    if (!wasEditing && isEditing) {
+      editFieldRef.current.focus();
+    }
+    if (wasEditing && !isEditing) {
+      editButtonRef.current.focus();
+    }
+  }, [wasEditing, isEditing]);
+
+  return <li className="todo">{isEditing ? editingTemplate : viewTemplate}</li>;
+}
