+++ src/index.css
@@ -50,7 +50,7 @@ button,
 input {
   overflow: visible;
 }
-input[type="text"] {
+input[type='text'] {
   border-radius: 0;
 }
 body {
@@ -69,7 +69,7 @@ body {
 }
 /*END RESETS*/
 /* GLOBAL STYLES */
-.form-group > input[type="text"] {
+.form-group > input[type='text'] {
   display: inline-block;
   margin-top: 0.4rem;
 }
@@ -83,7 +83,7 @@ body {
   border-width: 1px;
   border-color: #d3d3d3;
 }
-.btn.toggle-btn[aria-pressed="true"] {
+.btn.toggle-btn[aria-pressed='true'] {
   text-decoration: underline;
   border-color: #4d4d4d;
 }
@@ -123,7 +123,7 @@ body {
   clip: rect(1px, 1px, 1px, 1px);
   white-space: nowrap;
 }
-[class*="stack"] > * {
+[class*='stack'] > * {
   margin-top: 0;
   margin-bottom: 0;
 }
@@ -187,16 +187,16 @@ body {
   border-color: #4d4d4d;
   box-shadow: inset 0 0 0 2px;
 }
-[class*="__lg"] {
+[class*='__lg'] {
   display: inline-block;
   width: 100%;
   font-size: 1.9rem;
 }
-[class*="__lg"]:not(:last-child) {
+[class*='__lg']:not(:last-child) {
   margin-bottom: 1rem;
 }
 @media screen and (min-width: 620px) {
-  [class*="__lg"] {
+  [class*='__lg'] {
     font-size: 2.4rem;
   }
 }
@@ -237,14 +237,14 @@ body {
   clear: left;
 }
 .c-cb > label::before,
-.c-cb > input[type="checkbox"] {
+.c-cb > input[type='checkbox'] {
   box-sizing: border-box;
   top: -2px;
   left: -2px;
   width: 44px;
   height: 44px;
 }
-.c-cb > input[type="checkbox"] {
+.c-cb > input[type='checkbox'] {
   -webkit-font-smoothing: antialiased;
   cursor: pointer;
   position: absolute;
@@ -263,18 +263,18 @@ body {
   touch-action: manipulation;
 }
 .c-cb > label::before {
-  content: "";
+  content: '';
   position: absolute;
   border: 2px solid currentColor;
   background: transparent;
 }
-.c-cb > input[type="checkbox"]:focus + label::before {
+.c-cb > input[type='checkbox']:focus + label::before {
   border-width: 4px;
   outline: 3px dashed #228bec;
 }
 .c-cb > label::after {
   box-sizing: content-box;
-  content: "";
+  content: '';
   position: absolute;
   top: 11px;
   left: 9px;
@@ -287,6 +287,6 @@ body {
   opacity: 0;
   background: transparent;
 }
-.c-cb > input[type="checkbox"]:checked + label::after {
+.c-cb > input[type='checkbox']:checked + label::after {
   opacity: 1;
 }
