--- src/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-
-ReactDOM.render(<App />, document.getElementById("root"));
